livekit-agent/requirements.txt
livekit
livekit-agents[openai,elevenlabs,silero]
livekit-plugins-openai
livekit-plugins-elevenlabs
livekit-plugins-silero
python-dotenv
supabase
cohere
